⚰️ refactor: remove dead code related to whiteboard
excluding some that are still being used
