--- components/WhiteBoard/ObjectPropertyEditor/objectPropertyEditor.module.css
@@ -1,70 +1,0 @@
-.container {
-  padding: 15px 10px;
-  display: flex;
-  flex-direction: column;
-  gap: 5px;
-}
-
-.panel {
-  border-radius: 10px;
-  border: 1px solid var(--secondary-dark);
-  padding: 10px 5px;
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  gap: 5px;
-  flex: 0;
-}
-
-.property {
-  min-width: 0;
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-}
-
-.label {
-  min-width: 30;
-  font-size: 14px;
-  padding: 0;
-  margin: 0 8px;
-  text-align: center;
-  color: var(--secondary-dark);
-}
-
-.input {
-  text-align: center;
-  background-color: var(--background);
-  box-sizing: border-box;
-  border: 1px solid var(--background);
-  padding: 3px;
-  min-width: 0;
-  color: var(--secondary-light);
-  flex: 1;
-}
-
-.select {
-  text-align: center;
-  background-color: var(--background);
-  box-sizing: border-box;
-  border: 1px solid var(--background);
-  padding: 3px;
-  min-width: 0;
-  color: var(--secondary-light);
-  flex: 1;
-}
-
-.select:focus {
-  border-radius: 10px;
-  color: var(--secondary-dark);
-  background-color: var(--background);
-  border: 1px solid var(--secondary-dark);
-}
-
-.input:focus {
-  border-radius: 10px;
-  color: var(--secondary-dark);
-  background-color: var(--foreground);
-  border: 1px solid var(--secondary-dark);
-}
