--- components/WhiteBoard/TreeViewer/treeViewer.module.css
@@ -1,51 +1,0 @@
-.arrow {
-  font-size: 20px;
-  rotate: -90deg;
-  color: var(--secondary-light);
-  transition: rotate 200ms;
-  width: 20px;
-}
-
-.expand {
-  rotate: 0deg;
-}
-
-.nodeContainer {
-  cursor: default;
-  display: flex;
-  justify-content: flex-start;
-}
-
-.container {
-  padding: 10px 5px;
-}
-
-.objInfoContainer {
-  padding-left: 3px;
-  flex: 1;
-  cursor: pointer;
-  height: 24px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: 6px;
-}
-
-.objInfoContainer:hover {
-  border-radius: 10px;
-  background-color: var(--highlight);
-}
-
-.objIcon {
-  color: var(--secondary-light);
-  font-size: 20px;
-}
-
-.currentObj {
-  border-radius: 10px;
-  background-color: var(--foreground);
-}
-
-.nodeWithExpand {
-  margin-left: 15px;
-}
