--- states/whiteboard.ts
@@ -1,112 +1,0 @@
-import { constructRootObjTree, objNodeSorter, translateObj } from '@/utils/whiteboardHelper';
-import { create } from 'zustand';
-
-interface WhiteBoardState {
-  objMap: Map<string, Obj>;
-  graphData: GraphData | null;
-  objTree: ObjNode;
-  currentTool: Tool;
-  currentObj: string | null;
-  drag: DragData | null;
-  bundle: ObjBundle | null;
-}
-
-interface WhiteBoardActions {
-  addBundle: (offset: Coord) => void;
-  addObj: (obj: Obj, global?: boolean) => void;
-  updateObj: (obj: Obj) => void;
-  syncObjTree: () => void;
-  setCurrentTool: (tool: Tool) => void;
-  setCurrentObj: (obj: string | null) => void;
-  setDrag: (drag: DragData | null) => void;
-  loadDocument: (document: WBDocument) => void;
-  setBundle: (bundle: ObjBundle) => void;
-  removeObj: (obj: Obj, global?: boolean) => void;
-  resetWhiteBoard: () => void;
-  removeNode: (node: NodeData) => void;
-}
-
-const ROOT_OBJ = {
-  objId: 'root',
-  childNodes: [],
-  depth: 0,
-} as ObjNode;
-
-const initialState = {
-  objMap: new Map<string, Obj>(),
-  graphData: null,
-  objTree: ROOT_OBJ,
-  currentTool: 'SELECT',
-  currentObj: null,
-  drag: null,
-} as WhiteBoardState;
-
-export const useWhiteBoard = create<WhiteBoardState & WhiteBoardActions>()((set, get) => ({
-  ...initialState,
-  addObj: (obj: Obj, global: boolean = true) => {
-    set((state) => ({
-      objMap: state.objMap.set(obj.objId, obj),
-    }));
-    if (global) {
-      set((state) => ({
-        objTree: {
-          ...state.objTree,
-          childNodes: [
-            { objId: obj.objId, childNodes: [], depth: obj.depth },
-            ...state.objTree.childNodes,
-          ],
-        },
-        currentObj: obj.objId,
-      }));
-    }
-  },
-  updateObj: (obj: Obj) => set((state) => ({ objMap: state.objMap.set(obj.objId, obj) })),
-  syncObjTree: () => {
-    set({ objTree: constructRootObjTree(get().objMap) });
-  },
-  setCurrentTool: (tool: Tool) => set(() => ({ currentTool: tool })),
-  setCurrentObj: (obj: string | null) => set(() => ({ currentObj: obj })),
-  setDrag: (drag: DragData | null) => set(() => ({ drag: drag })),
-  loadDocument: async (document: WBDocument) => {
-    const newObjMap = new Map<string, Obj>(Object.entries(document.documentData));
-    const newObjTree = constructRootObjTree(newObjMap);
-    set(() => ({ objMap: newObjMap, objTree: newObjTree, graphData: document.graphData }));
-  },
-  setBundle: (bundle: ObjBundle) => set(() => ({ bundle: bundle })),
-  resetWhiteBoard: () => set(() => ({ ...initialState })),
-  addBundle: (offset: Coord) => {
-    const bundle = get().bundle;
-    if (bundle !== null) {
-      const realOffset = { x: -bundle.x + offset.x, y: -bundle.y + offset.y };
-      for (const obj of bundle.objs) {
-        get().addObj(translateObj(obj, realOffset));
-      }
-      // sort by depth
-      set((state) => ({
-        objTree: { ...state.objTree, childNodes: state.objTree.childNodes.sort(objNodeSorter) },
-      }));
-    }
-    set(() => ({
-      currentTool: 'SELECT',
-      bundle: null,
-    }));
-  },
-  removeObj: (obj: Obj, global: boolean = true) => {
-    set((state) => {
-      const map = new Map(state.objMap);
-      map.delete(obj.objId);
-      return { objMap: map };
-    });
-    if (global) {
-      // TODO: remove from objTree
-    }
-  },
-  removeNode: (node) => {
-    const data = get().graphData;
-    if (data === null) return;
-    //get index of node
-    const targetIndex = data.keywords.findIndex((value) => value === node.label);
-    if (targetIndex === -1) return;
-    // delete keyword;
-  },
-}));
