+++ utils/whiteboardHelper.ts
@@ -1,34 +1,4 @@
-import { useWhiteBoard } from '@/states/whiteboard';
-import {
-  forceCenter,
-  forceCollide,
-  forceLink,
-  forceManyBody,
-  forceSimulation,
-  hierarchy,
-  tree,
-} from 'd3';
 import { Camera, Vector2, Vector3 } from 'three';
-
-export const SELECT_HIGHLIGHT = '#9edae6';
-export const SELECT_HIGHLIGHT_OPACITY = 0;
-export const SELECT_DEPTH = 1;
-export const FRAME_DEPTH = 1.0001;
-export const FRAME_CORNER_DEPTH = 1.0002;
-export const FRAME_COLOR = '#3491e3';
-export const FRAME_WIDTH = 5;
-const MAX_DEPTH = 1;
-const MIN_DEPTH = 0;
-
-export const rootObj: Obj = {
-  objId: 'ROOT',
-  type: 'ROOT',
-  x: 0,
-  y: 0,
-  depth: MIN_DEPTH,
-  parentId: '',
-};
-
 export const DEFAULT_FONT = 'https://fonts.gstatic.com/ea/notosanskr/v2/NotoSansKR-Regular.woff';
 
 /**
@@ -41,35 +11,6 @@ export function genId(): string {
 }
 
 /**
- * Generate depth value between objA and objB
- *
- * @param {Obj} [objA] the object that stays below (optional) if left blank, uses MIN_DEPTH
- * @param {Obj} [objB] the object that stays above (optional) if left blank, uses MAX_DEPTH
- * @return {number} Generated depth value.
- */
-export function genDepth(objA?: Obj, objB?: Obj): number {
-  if (objA === undefined && objB === undefined) return topDepth();
-  if (objA === undefined) return MIN_DEPTH;
-  if (objB === undefined)
-    return objA.depth + 0.0001 < 1 ? objA.depth + 0.0001 : (objA.depth + MAX_DEPTH) / 2;
-  return (objA.depth + objB.depth) / 2;
-}
-
-/**
- * get depth value between top object and MAX_DEPTH
- *
- * @return {number} Generated depth value.
- */
-export function topDepth(): number {
-  const rootObjNode = useWhiteBoard.getState().objTree;
-  if (rootObjNode.childNodes.length === 0) {
-    return 0.0001;
-  }
-  const currentTop = rootObjNode.childNodes[0].depth || rootObjNode.depth;
-  return currentTop + 0.0001 < MAX_DEPTH ? currentTop + 0.0001 : (currentTop + MAX_DEPTH) / 2;
-}
-
-/**
  * Generate random color
  *
  * @return {string} Generated random color value in rgb string
@@ -78,124 +19,6 @@ export function genColor(): string {
   return `rgb(${Math.floor(Math.random() * 255)},${Math.floor(Math.random() * 255)},${Math.floor(
     Math.random() * 255,
   )})`;
-}
-
-/**
- * Create Rectangle Object based on input parameter
- *
- * @param {number} [x] x position
- * @param {number} [y] y position
- * @param {number} [w] width value
- * @param {number} [h] height value
- * @param {string} [color=genColor()] rgb string
- * @param {number} [depth=genDepth()] depth value
- * @return {RectObj} generated Rectangle Object based on input parameter
- */
-export function createRect(
-  x: number,
-  y: number,
-  w: number,
-  h: number,
-  color: string = genColor(),
-  depth: number = genDepth(),
-): RectObj {
-  // automatically validates
-  return validateRectObj({
-    objId: genId(),
-    type: 'RECT',
-    x: x,
-    y: y,
-    w: w,
-    h: h,
-    depth: depth,
-    color: color,
-    parentId: 'ROOT',
-  } as RectObj);
-}
-
-/**
- * Create Text Object based on input parameter
- *
- * @param {number} [x] x position
- * @param {number} [y] y position
- * @param {number} [w] width value
- * @param {number} [depth=genDepth()] depth value
- * @param {TextOptions} [options] other optional properties
- * @return {TextObj} generated Text Object based on input parameter
- */
-export function createText(
-  x: number,
-  y: number,
-  w: number,
-  depth: number = genDepth(),
-  options?: TextOptions,
-): TextObj {
-  return validateTextObj({
-    objId: genId(),
-    type: 'TEXT',
-    x: x,
-    y: y,
-    depth: depth,
-    parentId: 'ROOT',
-    w: w,
-    fontSize: options?.fontSize ?? 30,
-    overflow: options?.overflow ?? 'normal',
-    text: options?.text ?? '',
-    color: options?.color ?? genColor(),
-    textAlign: options?.textAlign ?? 'left',
-  } as TextObj);
-}
-
-export function createCircle(
-  x: number,
-  y: number,
-  r: number,
-  depth: number = genDepth(),
-  options?: CircleOptions,
-): CircleObj {
-  return validateCircleObj({
-    objId: genId(),
-    type: 'CIRCLE',
-    x: x,
-    y: y,
-    r: r,
-    depth: depth,
-    parentId: 'ROOT',
-    color: options?.color ?? genColor(),
-  } as CircleObj);
-}
-
-/**
- * Create Line Object based on input parameter
- *
- * @param {number} [x] x value
- * @param {number} [y] y value
- * @param {number} [x2] x2 value
- * @param {number} [y2] y2 value
- * @param {number} [depth=genDepth()] depth value
- * @param {LineOptions} [options] other options
- * @return {LineObj} generated Line Object based on input parameter
- */
-export function createLine(
-  x: number,
-  y: number,
-  x2: number,
-  y2: number,
-  depth: number = genDepth(),
-  options?: LineOptions,
-): LineObj {
-  return validateLineObj({
-    objId: genId(),
-    type: 'LINE',
-    x: x,
-    y: y,
-    x2: x2,
-    y2: y2,
-    strokeWidth: options?.strokeWidth ?? 3,
-    depth: depth,
-    parentId: 'ROOT',
-    color: options?.color ?? genColor(),
-  } as LineObj);
 }
 
 /**
@@ -213,51 +36,6 @@ export function getPos(mouse: Vector2, camera: Camera, validate: boolean = false
 }
 
 /**
- * Construct and return ObjTree based on ObjMap input
- *
- * @param {Map<string, Obj>} [objMap] Map of all objects
- * @return {ObjNode} Root node of constructed objTree
- */
-export function constructRootObjTree(objMap: Map<string, Obj>): ObjNode {
-  const dependencyMap = new Map<string, string[]>();
-  const blank: string[] = [];
-
-  for (const [_, v] of objMap) {
-    dependencyMap.set(v.parentId, [...(dependencyMap.get(v.parentId) || blank), v.objId]);
-  }
-
-  const root: ObjNode = {
-    objId: 'ROOT',
-    childNodes: [],
-    depth: 0,
-  };
-
-  const stack: ObjNode[] = [root];
-
-  while (stack.length > 0) {
-    const parent = stack.pop()!;
-    const nextIds = dependencyMap.get(parent.objId);
-    if (nextIds === undefined) continue;
-    for (const id of nextIds) {
-      const obj = objMap.get(id);
-      if (obj === undefined) continue;
-      const child: ObjNode = { objId: id, childNodes: [], depth: obj.depth };
-      parent.childNodes.push(child);
-      stack.push(child);
-    }
-    // sort obj by depth
-    parent.childNodes.sort(objNodeSorter);
-  }
-
-  return root;
-}
-
-export function objNodeSorter(a: ObjNode, b: ObjNode) {
-  if (a.depth < b.depth) return 1;
-  return -1;
-}
-
-/**
  * returns number value in provided boundary
  *
  * @param {number} [value] value to bound
@@ -267,74 +45,6 @@ export function objNodeSorter(a: ObjNode, b: ObjNode) {
  */
 export function boundNumber(value: number, min: number, max: number): number {
   return Math.max(Math.min(value, max), min);
-}
-
-/**
- * validates input obj
- *
- * @param {Obj} [obj] obj to validate
- * @return {Obj} validated obj
- */
-export function validateObj(obj: Obj): Obj {
-  switch (obj.type) {
-    case 'RECT':
-      validateRectObj(obj as RectObj);
-      break;
-    case 'ROOT':
-      break;
-    case 'TEXT':
-      validateTextObj(obj as TextObj);
-      break;
-  }
-  return obj;
-}
-
-/**
- * validates input rect obj
- *
- * @param {RectObj} [obj] rect obj to validate
- * @return {Obj} validated obj
- */
-function validateRectObj(obj: RectObj): RectObj {
-  obj.w = validateValue(obj.w);
-  obj.h = validateValue(obj.h);
-  obj.x = validateValue(obj.x);
-  obj.y = validateValue(obj.y);
-  return obj;
-}
-
-/**
- * validates input line obj
- *
- * @param {LineObj} [obj] line obj to validate
- * @return {LineObj} validated obj
- */
-function validateLineObj(obj: LineObj): LineObj {
-  obj.x2 = validateValue(obj.x2);
-  obj.y2 = validateValue(obj.y2);
-  obj.x = validateValue(obj.x);
-  obj.y = validateValue(obj.y);
-  return obj;
-}
-
-function validateCircleObj(obj: CircleObj): CircleObj {
-  obj.x = validateValue(obj.x);
-  obj.y = validateValue(obj.y);
-  obj.r = validateValue(obj.r);
-  return obj;
-}
-
-/**
- * validates input text obj
- *
- * @param {TextObj} [obj] text obj to validate
- * @return {TextObj} validated obj
- */
-function validateTextObj(obj: TextObj): TextObj {
-  obj.w = validateValue(obj.w);
-  obj.x = validateValue(obj.x);
-  obj.y = validateValue(obj.y);
-  return obj;
 }
 
 /**
@@ -350,306 +60,10 @@ export function validateValue(value: number, positive = false): number {
   return newValue;
 }
 
-interface NodeData {
-  id: number;
-  label: string;
-  x?: number;
-  y?: number;
-  vx?: number;
-  vy?: number;
-}
-
-interface LinkData {
-  source: number;
-  target: number;
-}
-
-export function createTreeFromMindmap(response: GraphData): ObjBundle {
-  let top = topDepth();
-  const fontSize = 30;
-  const edgeColor = '#dddddd';
-  const fontColor = '#000000';
-  const nodeColor = '#000000';
-  const unit = 0.0001;
-  const result: Obj[] = [];
-  const bounds = { minX: NaN, maxX: NaN, minY: NaN, maxY: NaN };
-
-  const updateBounds = (obj: Obj) => {
-    if (isNaN(bounds.minX)) {
-      bounds.minX = obj.x;
-      bounds.maxX = obj.x;
-      bounds.minY = obj.y;
-      bounds.maxY = obj.y;
-      return;
-    }
-    bounds.minX = Math.min(bounds.minX, obj.x);
-    bounds.maxX = Math.max(bounds.maxX, obj.x);
-    bounds.minY = Math.min(bounds.minY, obj.y);
-    bounds.maxY = Math.max(bounds.maxY, obj.y);
-  };
-
-  const appendObj = (obj: Obj) => {
-    updateBounds(obj);
-    result.push(obj);
-    top += unit;
-  };
-
-  const graph = new Map<string, number[]>(Object.entries(response.graph));
-  const root = hierarchy(response.root as unknown, (d) => {
-    return graph.get((d as number).toString());
-  });
-
-  const pointRadius = 5;
-
-  const layout = tree().nodeSize([60, 300]);
-  const data = layout(root);
-
-  data.links().map((v) => {
-    appendObj(
-      createLine(v.source.y, v.source.x, v.target.y, v.target.x, top, { color: edgeColor }),
-    );
-  });
-  data.descendants().map((v) => {
-    appendObj(
-      createText(v.y - 50, v.x + pointRadius * 2 + 5, 100, top, {
-        color: fontColor,
-        fontSize: fontSize,
-        text: response.keywords[v.data as number],
-        textAlign: 'center',
-      }),
-    );
-    appendObj(
-      createCircle(v.y - pointRadius, v.x - pointRadius, pointRadius, top, { color: nodeColor }),
-    );
-  });
-  return {
-    x: validateValue((bounds.maxX + bounds.minX) / 2),
-    y: validateValue((bounds.maxY + bounds.minY) / 2),
-    w: bounds.maxX - bounds.minX,
-    h: bounds.maxY - bounds.minY,
-    objs: result,
-  };
-}
-
-export function createForceBundleFromMindmap(response: GraphData): ObjBundle {
-  const fontSize = 20;
-  const radius = 50;
-  const edgeWidth = 2;
-  const edgeColor = '#31493C';
-  const fontColor = '#FFFFFF';
-  const nodeColor = '#7A9E7E';
-  const bounds = { minX: NaN, maxX: NaN, minY: NaN, maxY: NaN };
-  let top = topDepth();
-  const unit = 0.0001;
-  const result: Obj[] = [];
-
-  const updateBounds = (obj: Obj) => {
-    if (isNaN(bounds.minX)) {
-      bounds.minX = obj.x;
-      bounds.maxX = obj.x;
-      bounds.minY = obj.y;
-      bounds.maxY = obj.y;
-      return;
-    }
-    bounds.minX = Math.min(bounds.minX, obj.x);
-    bounds.maxX = Math.max(bounds.maxX, obj.x);
-    if (obj.type === 'TEXT') {
-      bounds.maxX = Math.max(bounds.maxX, obj.x + (obj as TextObj).w);
-    }
-    bounds.minY = Math.min(bounds.minY, obj.y);
-    bounds.maxY = Math.max(bounds.maxY, obj.y + fontSize);
-  };
-
-  const appendObj = (obj: Obj) => {
-    updateBounds(obj);
-    result.push(obj);
-    top += unit;
-  };
-
-  const simulation = forceSimulation<NodeData>();
-
-  const nodes: NodeData[] = [];
-
-  response.keywords.forEach((v, i) => {
-    nodes.push({ id: i, label: v });
-  });
-
-  const links: any[] = [];
-
-  const graph = new Map<string, number[]>(Object.entries(response.graph));
-  for (const entry of graph.entries()) {
-    for (const child of entry[1]) {
-      links.push({ source: parseInt(entry[0]), target: child });
-    }
-  }
-
-  simulation
-    .nodes(nodes)
-    .force(
-      'link',
-      forceLink<NodeData, LinkData>(links)
-        .id((d) => d.id)
-        .strength(2),
-    )
-    .force('charge', forceManyBody<NodeData>().strength(-100))
-    .force('center', forceCenter(0, 0).strength(1))
-    .force('collision', forceCollide(radius * 2))
-    .tick(300);
-
-  for (const link of links) {
-    const source = link.source as NodeData;
-    const target = link.target as NodeData;
-    appendObj(
-      createLine(source.x || 0, source.y || 0, target.x || 0, target.y || 0, top, {
-        strokeWidth: edgeWidth,
-        color: edgeColor,
-      }),
-    );
-  }
-
-  for (const node of nodes) {
-    appendObj(
-      createCircle(
-        (node.x ?? 0) - radius,
-        (node.y ?? 0) - radius,
-        radius,
-
-        top,
-        { color: nodeColor },
-      ),
-    );
-    appendObj(
-      createText((node.x ?? 0) - radius, (node.y ?? 0) - fontSize / 2, radius * 2, top, {
-        textAlign: 'center',
-        fontSize: fontSize,
-        text: node.label,
-        color: fontColor,
-        overflow: 'break-word',
-      }),
-    );
-  }
-
-  return {
-    x: validateValue((bounds.maxX + bounds.minX) / 2),
-    y: validateValue((bounds.maxY + bounds.minY) / 2),
-    w: bounds.maxX - bounds.minX,
-    h: bounds.maxY - bounds.minY,
-    objs: result,
-  };
-}
-
 export function calculateLineGeometry(x: number, y: number, x2: number, y2: number) {
   const w = Math.sqrt(Math.pow(x - x2, 2) + Math.pow(y - y2, 2));
   const d = Math.atan2(y2 - y, x2 - x);
   return { w, d };
-}
-
-// DAGRE.js tree version
-//
-// export function createTreeBundleFromMindmap(response: MindmapResponse): ObjBundle {
-//   const fontSize = 20;
-//   const bounds = { minX: NaN, maxX: NaN, minY: NaN, maxY: NaN };
-//   let top = topDepth();
-//   const unit = 0.0001;
-//   const result: Obj[] = [];
-
-//   const updateBounds = (obj: Obj) => {
-//     if (isNaN(bounds.minX)) {
-//       bounds.minX = obj.x;
-//       bounds.maxX = obj.x;
-//       bounds.minY = obj.y;
-//       bounds.maxY = obj.y;
-//       return;
-//     }
-//     bounds.minX = Math.min(bounds.minX, obj.x);
-//     bounds.maxX = Math.max(bounds.maxX, obj.x);
-//     if (obj.type === 'TEXT') {
-//       bounds.maxX = Math.max(bounds.maxX, obj.x + (obj as TextObj).w);
-//     }
-//     bounds.minY = Math.min(bounds.minY, obj.y);
-//     bounds.maxY = Math.max(bounds.maxY, obj.y + fontSize);
-//   };
-
-//   const appendObj = (obj: Obj) => {
-//     updateBounds(obj);
-//     result.push(obj);
-//     top += unit;
-//   };
-
-//   const g = new graphlib.Graph()
-//     .setGraph({
-//       rankdir: 'LR',
-//       edgesep: 100,
-//       ranksep: 50,
-//       nodesep: 30,
-//     })
-//     .setDefaultEdgeLabel(() => ({}));
-
-//   response.keywords.forEach((v, i) => {
-//     g.setNode(i.toString(), { label: v, width: fontSize * v.length, height: 30 });
-//   });
-
-//   const graph = new Map<string, number[]>(Object.entries(response.graph));
-//   for (const entry of graph.entries()) {
-//     for (const child of entry[1]) {
-//       g.setEdge(entry[0].toString(), child.toString());
-//     }
-//   }
-
-//   dagre.layout(g);
-
-//   g.edges().forEach((v) => {
-//     const edge = g.edge(v);
-//     for (let i = 1; i < edge.points.length; i++) {
-//       appendObj(
-//         createLine(
-//           edge.points[i].x,
-//           edge.points[i].y,
-//           edge.points[i - 1].x,
-//           edge.points[i - 1].y,
-//           top,
-//         ),
-//       );
-//     }
-//   });
-
-//   g.nodes().forEach((v) => {
-//     const node = g.node(v);
-//     appendObj(
-//       createText(node.x - node.width / 2, node.y - node.height / 2, node.width, top, {
-//         fontSize: fontSize,
-//         textAlign: 'center',
-//         color: 'black',
-//         text: node.label,
-//       }),
-//     );
-//   });
-
-//   const output = g.graph();
-//   if (output.width === undefined || output.height === undefined)
-//     throw new Error('그래프 생성에 실패하였습니다.');
-
-//   return {
-//     x: output.width / 2,
-//     y: output.height / 2,
-//     w: output.width,
-//     h: output.height,
-//     objs: result,
-//   };
-// }
-
-export function translateObj(obj: Obj, offset: Coord): Obj {
-  switch (obj.type) {
-    case 'LINE':
-      (obj as LineObj).x2 = validateValue((obj as LineObj).x2 + offset.x);
-      (obj as LineObj).y2 = validateValue((obj as LineObj).y2 + offset.y);
-    case 'RECT':
-    case 'TEXT':
-    case 'CIRCLE':
-      obj.x = validateValue(obj.x + offset.x);
-      obj.y = validateValue(obj.y + offset.y);
-  }
-  return obj;
 }
 
 export const lipsum = `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur pulvinar, nulla quis viverra venenatis, metus sapien blandit urna, nec tristique sem justo non justo. Pellentesque semper massa nec dapibus luctus. Vestibulum facilisis ornare augue vel semper. Pellentesque id faucibus augue. Quisque ullamcorper tempor magna eget molestie. Etiam mattis a velit quis porttitor. Sed et posuere sapien, non convallis elit. Mauris tempor, metus non auctor accumsan, ante lacus posuere augue, ac scelerisque sem nunc luctus arcu. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia curae;
